Implement generalized construction from cell types.
This change introduces the `FromCell` trait, which is used to allow more
generalized construction of `Content` types from cells. A cell is
similar to a grapheme or code point, but can be represented by types
that included additional information. Of note is `StyledCell`, which
allows a grapheme to be paired with a style to construct a `Styled`
while maintaining that the constructed `Styled` represents a unitary
character.