Implement more sophisticated unit types.
This change introduces lower level APIs that model Unicode graphemes and
a bespoke concept of a morpheme. Morphemes are a subset of Unicode
graphemes supported by the library, and are most concerned with
separating graphemes into narrow (or ambiguous) and wide base loosely on
Unicode Annex #11. This should bring API-level support for text
comprised entirely of graphemes with a column width (unrelated to
glyphs, fonts, and rendering) of either one or two.

Atop these APIs are a `Text` type, which is a UTF-8 string (borrowed or
owned) comprised of morphemes. `Segment` pairs `Text` with an instance
of some other type (by default the unit type) intended for styling and
removes any line breaks when constructed. `Line` is formed from an
ordered sequence of `Segment`s. These types can form the basis for
`Block`.