Remove error branching for one-cell width morphemes.
This change introduces the `Congruence` trait, which is a supertrait of
`MorphemeKind`. This trait allows morpheme kinds to specify an error
type and a `congruence` function that determines if a morpheme is
congruent with a given width (can represent that width). For morphemes
that support one-cell widths, this is infallible. By specifying
`Infallible` as the error type, these morphemes no longer require error
branching when constructing block text types from widths.