Implement detection of the output environment.
This change ports detection of encodings, width, etc. from `pusheen`.
`RenderContext` now interacts with this, allowing rendering to react to
the configured output environment. ANSI styles are now applied based on
this information, though this is somewhat incomplete. Lossy conversion
of Unicode to ASCII outputs has not yet been implemented in this change.