Pad lines with a right margin with ephemeral segments.
This change introduces a `BlockWidth` render node and uses ephemeral
`BlankSegment`s and graphemes in `BlockText` and `Render`
implementations to pad lines as needed. Importantly, `Block`s no longer
append `BlankSegment`s to their lines for this purpose. This complicates
rendering, but with a very useful mechanism. This also avoids the need
to scan and pad lines whenever a `Block` is modified, which precludes
some features, can be costly, and is an invariant that is difficult to
maintain.